Big overhaul
I've been working with ChatGPT and Grok a lot. This commit is an
attempt to clean up, and settle on some decisions like which
part of the state should be int vs float, optional, etc.

+++ src/main/Constants.h
@@ -1,8 +1,7 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 
 static const double INITIAL_SAMPLES_PER_PIXEL = 1;
 static const double INITIAL_VERTICAL_ZOOM = 1;
-static const int64_t INITIAL_SAMPLE_OFFSET = 0;
-
+static const size_t INITIAL_SAMPLE_OFFSET = 0;
